requirements.txt
modules==1.0.0
calculator==1.1.0
future==0.16.0
unidecode==0.04.21
selenium==3.6.0
bs4==0.0.1
pytesseract==0.1.7
Pillow==4.3.0
opencv-python==3.3.0.10
simplejson==3.11.1
termcolor==1.1.0
